(cleanup) Refactor BottomNavBar and Splash folders
