+++ client/public/styles/skeleton.css
@@ -409,7 +409,7 @@ form {
   height: 50vh;
   background-color: #29D9C2;
 }
-.CatchARideButton {
+.RideButton {
   height: 50vh;
   background-color: #BDF271;
 }
